Pull out TJTwitterStatistics
