+++ src/main/java/edu/woodson/TJTwitter.java
@@ -17,8 +17,7 @@ import java.util.stream.StreamSupport;
 
 class TJTwitter {
     private static final String COMMON_WORDS_LOCATION = "/commonWords.txt";
-    private final List<Status> statuses = new ArrayList<>();
-    private final List<String> words = new ArrayList<>();
+    private final TJTwitterStatistics statistics = new TJTwitterStatistics();
     private final Twitter twitter;
 
     public TJTwitter(Twitter twitter) {
@@ -26,57 +25,13 @@ class TJTwitter {
     }
 
     public int getMaxFrequency() {
-        return Math.toIntExact(calculateMax(createFrequencyMap(words)).orElseThrow());
-    }
-
-    Optional<Long> calculateMax(Map<String, Long> map) {
-        return map.values().stream().max(Long::compareTo);
-    }
-
-    /******************  Part III - Test *******************/
-
-    Map<String, Long> createFrequencyMap(List<String> words) {
-        return words.stream().collect(Collectors.toMap(Function.identity(), string -> words.stream()
-                .filter(s -> s.equals(string))
-                .count(), Long::max));
+        return statistics.getMaxFrequency();
     }
 
     /******************  Part III - Tweet *******************/
 
     public String getMostPopularWord() {
-        return CollectionUtil.toSingle(mostPopularWord(words));
-    }
-
-    /**
-     * This method calculates the word that appears the most times
-     * Consider case - should it be case sensitive?  The choice is yours.
-     * <p>
-     * post will populate the maxFrequency variable with the frequency of the most common word
-     *
-     * @param words The words to search.
-     * @post will populate the maxFrequency variable with the frequency of the most common word
-     */
-    @SuppressWarnings("unchecked")
-    public Set<String> mostPopularWord(List<String> words) {
-        if (words.isEmpty()) {
-            throw new IllegalArgumentException("No words found for " + words);
-        }
-
-        Map<String, Long> map = createFrequencyMap(words);
-        long max = calculateMax(words, map);
-        return map.keySet()
-                .stream()
-                .filter(s -> map.get(s).equals(max))
-                .collect(Collectors.toSet());
-    }
-
-    long calculateMax(List<String> words, Map<String, Long> map) {
-        Optional<Long> maxOptional = calculateMax(map);
-        if (!maxOptional.isPresent()) {
-            throw new IllegalArgumentException("No maximum found for " + words);
-        }
-
-        return maxOptional.get();
+        return statistics.getMostPopularWord();
     }
 
     /******************  Part IV *******************/
@@ -92,14 +47,12 @@ class TJTwitter {
      */
     @SuppressWarnings("unchecked")
     public void queryHandle(String handle) throws IOException {
-        statuses.clear();
-        statuses.addAll(fetchTweets(handle));
+        List<Status> statuses = fetchTweets(handle);
+        List<String> words = splitIntoWords(toMessage(statistics.getStatuses()));
+        statistics.setValues(statuses, words);
 
-        words.clear();
-        words.addAll(splitIntoWords(toMessage(statuses)));
-
-        this.words.removeAll(loadCommonWordsFromLocation());
-        this.words.addAll(sortAndRemoveEmpties(words));
+        statistics.removeCommonWords(loadCommonWordsFromLocation());
+        statistics.sortAndRemoveEntries();
     }
 
     /**
@@ -156,18 +109,6 @@ class TJTwitter {
         return loadCommonWordsFromStream(commonWordsURL.get().openStream());
     }
 
-    /**
-     * This method sorts the words in words in alphabetically (and lexicographic) order.
-     * You should use your sorting code you wrote earlier this year.
-     * Remove all empty strings while you are at it.
-     *
-     * @param words The words.
-     */
-    @SuppressWarnings("unchecked")
-    public List<String> sortAndRemoveEmpties(List<String> words) {
-        Collections.sort(removeEmptyStrings(words));
-        return words;
-    }
 
     Optional<URL> getCommonWordsURL() {
         return Optional.ofNullable(getClass().getResource(TJTwitter.COMMON_WORDS_LOCATION));
@@ -184,12 +125,6 @@ class TJTwitter {
                 .collect(Collectors.toList());
     }
 
-    List<String> removeEmptyStrings(List<String> terms) {
-        return terms.stream()
-                .map(String::trim)
-                .filter(s -> s.length() != 0)
-                .collect(Collectors.toList());
-    }
 
     /**
      * This method removes common punctuation from each individual word.
